Query implementations for storage server
Code for the Java application that will run on the storage server. It is possible to query the server via a socket to retrieve data. Query format is fetch;country1,country2,country3,etc;count where count is the rows of data per country sorted by windchill. Alternatively 'update' can be called to fetch the 10 coldest places for France, United States, Mexico and Spain.